agent_T2SQL_langchain/pyproject.toml
[project]
name = "a2a-t2sql-agent-langgraph"
version = "0.1.0"
description = "t2sql's LangGraph agent with A2A Protocol"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk==0.2.6",
    "httpx>=0.28.1",
    "langchain-google-genai>=2.0.10",
    "langgraph>=0.3.18",
    "pydantic>=2.10.6",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
    "langchain-core",
    "langchain-community>=0.3.12",
    "pydantic",
    "pymysql",
    "cryptography"
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
